divorce-prediction-system/requirements.txt
django-crispy-forms
crispy-bootstrap5
pandas
numpy
scikit-learn
pillow